+++ yandex/data_source_yandex_function_test.go
@@ -85,6 +85,12 @@ func TestAccDataSourceYandexFunction_full(t *testing.T) {
 	params.envKey = "tf_function_env"
 	params.envValue = "tf_function_env_value"
 	params.tags = acctest.RandomWithPrefix("tf-function-tag")
+	params.secret = testSecretParameters{
+		secretName:   "tf-function-secret-name",
+		secretKey:    "tf-function-secret-key",
+		secretEnvVar: "TF_FUNCTION_ENV_KEY",
+		secretValue:  "tf-function-secret-value",
+	}
 	params.zipFilename = "test-fixtures/serverless/main.zip"
 
 	resource.Test(t, resource.TestCase{
@@ -109,6 +115,10 @@ func TestAccDataSourceYandexFunction_full(t *testing.T) {
 					resource.TestCheckResourceAttrSet(functionDataSource, "version"),
 					resource.TestCheckResourceAttrSet(functionDataSource, "image_size"),
 					resource.TestCheckResourceAttrSet(functionDataSource, "loggroup_id"),
+					resource.TestCheckResourceAttrSet(functionDataSource, "secrets.0.id"),
+					resource.TestCheckResourceAttrSet(functionDataSource, "secrets.0.version_id"),
+					resource.TestCheckResourceAttr(functionDataSource, "secrets.0.key", params.secret.secretKey),
+					resource.TestCheckResourceAttr(functionDataSource, "secrets.0.environment_variable", params.secret.secretEnvVar),
 					testAccCheckCreatedAtAttr(functionDataSource),
 				),
 			},
@@ -179,6 +189,12 @@ resource "yandex_function" "test-function" {
     %s = "%s"
   }
   tags = ["%s"]
+  secrets {
+    id = yandex_lockbox_secret.secret.id
+    version_id = yandex_lockbox_secret_version.secret_version.id
+    key = "%s"
+    environment_variable = "%s"
+  }
   content {
     zip_filename = "%s"
   }
@@ -186,6 +202,26 @@ resource "yandex_function" "test-function" {
 
 resource "yandex_iam_service_account" "test-account" {
   name = "%s"
+}
+
+resource "yandex_resourcemanager_folder_iam_binding" "admin-account-iam" {
+  folder_id   = yandex_lockbox_secret.secret.folder_id
+  role        = "lockbox.payloadViewer"
+  members     = [
+    "serviceAccount:${yandex_iam_service_account.test-account.id}",
+  ]
+}
+
+resource "yandex_lockbox_secret" "secret" {
+  name        = "%s"
+}
+
+resource "yandex_lockbox_secret_version" "secret_version" {
+  secret_id = yandex_lockbox_secret.secret.id
+  entries {
+    key        = "%s"
+    text_value = "%s"
+  }
 }
 	`,
 		params.name,
@@ -199,6 +235,11 @@ resource "yandex_iam_service_account" "test-account" {
 		params.envKey,
 		params.envValue,
 		params.tags,
+		params.secret.secretKey,
+		params.secret.secretEnvVar,
 		params.zipFilename,
-		params.serviceAccount)
+		params.serviceAccount,
+		params.secret.secretName,
+		params.secret.secretKey,
+		params.secret.secretValue)
 }