+++ CHANGELOG.md
@@ -19,6 +19,7 @@ FEATURES:
 * **New Data Source:** `yandex_lockbox_secret_version`
 * **New Resource:** `yandex_lockbox_secret`
 * **New Resource:** `yandex_lockbox_secret_version`
+* serverless: add `secrets` attribute in `yandex_function` resource and data source
 
 ## 0.80.0 (September 22, 2022)
 BUG FIXES: